add CRUD comment

+++ models/comments.go
@@ -5,9 +5,9 @@ import "time"
 type Comments struct {
 	ID         uint64    `gorm:"primaryKey;autoIncrement" json:"id"`
 	Comment    string    `gorm:"type:text" json:"comment"`
-	CourseID   uint      `gorm:"not null" json:"-"`
+	CourseID   uint64    `gorm:"not null" json:"-"`
 	Course     Courses   `gorm:"foreignKey:CourseID;constraint:onUpdate:CASCADE,onDelete:CASCADE" json:"course_id"`
-	LecturerID uint      `gorm:"not null" json:"-"`
+	LecturerID uint64    `gorm:"not null" json:"-"`
 	Lecturer   Lecturers `gorm:"foreignKey:LecturerID;constraint:onUpdate:CASCADE,onDelete:CASCADE" json:"lecturer_id"`
 	SchoolYear uint      `gorm:"not null" json:"school_year"`
 	Semester   uint      `gorm:"not null" json:"semester"`
